pyproject.toml
[project]
name = "rrl_mujococ_cw2"
version = "1.0.0"
description = "Course Work 2 for RRL course - SAC and PPO"
authors = [{ name = "Leonard Hinckeldey", email = "l.hinckeldey@ed.ac.uk" }]
requires-python = ">=3.10,<3.11"
readme = "README.md"
license = "MIT"
keywords = [
    "reinforcement learning",
    "machine learning",
    "research",
]
dependencies = [
    "tensorboard>=2.10.0",
    "wandb>=0.13.11",
    "gym==0.23.1",
    "gymnasium[mujoco]==0.29.1",
    "torch==2.4.1",
    "moviepy>=1.0.3",
    "pygame>=2.1",
    "huggingface-hub>=0.11.1",
    "rich<12.0",
    "tenacity>=8.2.2",
    "tyro>=0.5.10",
    "pyyaml>=6.0",
    "mujoco<=2.3.3",
    "imageio>=2.14.1,<3",
    "notebook>=7.1.3",
    "matplotlib>=3.10.0",
]

[project.optional-dependencies]
mujoco = [
    "mujoco<=2.3.3",
    "imageio>=2.14.1,<3",
]
plot = []
pytest = ["pytest>=7.1.3,<8"]

[dependency-groups]
dev = ["pre-commit>=2.20.0,<3"]

[tool.hatch.build.targets.sdist]
include = [
    "agents",
]

[tool.hatch.build.targets.wheel]
include = [
    "agents",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
